requirements.txt
click
python-binance
python-dotenv
fastapi
uvicorn
jinja2
python-multipart
pytest
